decimation: decimate the audio path by 4
Decimate the audio path for processing down from the raw
44kHz to 11kHz. The main goal here is not to reduce the processing
overheads (although it does drop a 'full' SSB NR compute from 20%
down to 7%), but is to allow the noise reductions that use 'bins' to
focus those bins on the audio spectrom (0-5.5kHz), rather than spread
them across the full CD quality bandwidth (0-24kHz). We *could* have
bumped the number of bins in the NR systems (from 128 iirc) to get the
same effect maybe, but then that would have cost us more CPU time for
the same gain??

The change seems to help. Now we need to see if it helps when combined
with the FIR bandpass normalisation code...

Fixes: #2

Signed-off-by: Graham Whaley <graham.whaley@gmail.com>