schematic: fix wrong I2C bus wiring
Oops, the schematic had the LCD wired to the second (xxx1) I2C bus, but
actually it should be wired to the first (xxx0) - shared with the audio
board - as that is the 'default' arduino I2C bus for the Teensy 4.0.

Whislt editing the schematic, also add a warning about using external
PSU and plugging in USB - you need to snip a track on the Teensy. See
the Teensy docs for more details.

Signed-off-by: Graham Whaley <graham.whaley@gmail.com>