k4icy decoder: improve 'failed to decode' char
We had set the 'failed to decode' char to a '?', but of course
that is a valid char, so is not actually that nice on the screen.
Looking at the 2x16 char map, the 'solid block' char of 0xff is
a better thing to draw - more obvious it is a 'failed decode'.

Signed-off-by: Graham Whaley <graham.whaley@gmail.com>