cw: do not update tuning 'indicator' when in menu
We constantly assess and update the CW 'tuning' indicator
(<< < x > >>) when in a CW mode. This constantly calls the
lcd.createChar() function, which was corrupting the lcd
cursor position whilst in the menu system. This made it
difficult to enter slot names for instance. This is
'correct' behaviour from the LCD library, as creating a
char in CGRAM has to update the address counter (AC), to
write the CGRAM, which is the same AC that is used to draw
the cursor position.

Fix by never creating a char whilst the menu is active.

Fixes: #7

Signed-off-by: Graham Whaley <graham.whaley@gmail.com>