settings: name field: stylistic cleanups
There was a report that saving the name into a slot did not work. Having
re-examined the code, although it was not obvious how it worked, it should
have been working, and worked in my tests. The code could do with some minor
fixups though, such as not hard coding the field length of '4' - as at least
one user has modified their code to make the name fields longer, and some
of the code may not have handled that well.

Signed-off-by: Graham Whaley <graham.whaley@gmail.com>